fix(mobile): move modal overlay to profile and update imports
